+++ brainbuilder/utils/sonata/convert.py
@@ -258,6 +258,11 @@ def write_node_set_from_targets(target_files, output_file, cells_path):
         if re_layer.match(target_name):
             return {"layer": str(re_layer.match(target_name).group(1))}
         if re_node_id.match(target_content[0]):
+            L.warning(
+                "Note: a list of `node_id`s are being created, "
+                "the %s node_set should have a population added",
+                target_name,
+            )
             # targets are built from a mvd3 file so indexing starts from 1 compare to 0 in SONATA
             return {"node_id": (cells.ids(target_name) - 1).tolist()}